chore(*): chore all strange and useless things from project
Closes #15

Date:      Tue Jan 14 19:58:53 2021 +0700
