--- src/main/java/bp/constraint/AtLeastOneFieldNotNull.java
@@ -1,42 +1,0 @@
-package bp.constraint;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * AtLeastOneFieldNotNull annotation
- *
- * @author  Nikolai Osipov <nao99.dev@gmail.com>
- * @version 1.0.0
- * @since   2020-12-20
- */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Constraint(validatedBy = AtLeastOneFieldNotNullValidator.class)
-public @interface AtLeastOneFieldNotNull {
-    /**
-     * Fields
-     */
-    String[] fields();
-
-    /**
-     * Error message
-     */
-    String message() default "At least field should be not null";
-
-    /**
-     * Groups
-     */
-    Class<?>[] groups() default {};
-
-    /**
-     * Payload
-     */
-    Class<? extends Payload>[] payload() default {};
-}
