--- src/main/java/bp/exception/section/SectionException.java
@@ -1,47 +1,0 @@
-package bp.exception.section;
-
-import org.springframework.lang.NonNull;
-
-/**
- * SectionException class
- *
- * @author  Nikolai Osipov <nao99.dev@gmail.com>
- * @version 1.0.0
- * @since   2020-12-19
- */
-public class SectionException extends Exception {
-    /**
-     * SectionException constructor
-     */
-    public SectionException() {
-        super();
-    }
-
-    /**
-     * SectionException constructor
-     *
-     * @param message an exception message
-     */
-    public SectionException(@NonNull String message) {
-        super(message);
-    }
-
-    /**
-     * SectionException constructor
-     *
-     * @param message an exception message
-     * @param cause   a previous exception
-     */
-    public SectionException(@NonNull String message, @NonNull Throwable cause) {
-        super(message, cause);
-    }
-
-    /**
-     * SectionException constructor
-     *
-     * @param cause a previous exception
-     */
-    public SectionException(@NonNull Throwable cause) {
-        super(cause);
-    }
-}
