--- src/main/java/bp/repository/CourseRepository.java
@@ -1,16 +1,0 @@
-package bp.repository;
-
-import bp.entity.Course;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * CourseRepository interface
- *
- * @author  Nikolai Osipov <nao99.dev@gmail.com>
- * @version 1.0.0
- * @since   2020-12-16
- */
-@Repository
-public interface CourseRepository extends CrudRepository<Course, Long> {
-}
