+++ src/main/resources/db/migration/V1__Create_courses_table.sql
@@ -3,7 +3,6 @@ CREATE TABLE courses (
     name        VARCHAR (255) NOT NULL,
     description TEXT DEFAULT NULL,
     created_at  TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP(0),
-    updated_at  TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP(0),
 
     CONSTRAINT pk_courses__id PRIMARY KEY (id)
 );
